% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_two.R
\name{subtract_two}
\alias{subtract_two}
\title{Title}
\usage{
subtract_two(first_number, second_number)
}
\arguments{
\item{first_number}{The larger number we're subtracting from}

\item{second_number}{The smaller number we're subtracting}
}
\value{
The subtraction of first_number - second_number
}
\description{
Title
}
\examples{
subtract_two(10, 6)
}
